requirements.txt
fastapi
uvicorn
langchain_google_genai
langchain_core.messages
